+++ TaskLayer/XLSearchTask/XLSearchTask2.cs
@@ -148,8 +148,8 @@ namespace TaskLayer
             _pepxml.summary_xml = items[0].FullFilePath + ".pep.xml";              
 
             string proteaseC = ""; string proteaseNC = "";
-            foreach (var x in CommonParameters.Protease.SequencesInducingCleavage) { proteaseC += x; }
-            foreach (var x in CommonParameters.Protease.SequencesPreventingCleavage) { proteaseNC += x; }
+            foreach (var x in CommonParameters.DigestionParams.Protease.SequencesInducingCleavage) { proteaseC += x; }
+            foreach (var x in CommonParameters.DigestionParams.Protease.SequencesPreventingCleavage) { proteaseNC += x; }
 
             CrosslinkerTypeClass crosslinker = new CrosslinkerTypeClass().SelectCrosslinker(XlSearchParameters.CrosslinkerType);
             string modsFixed = ""; string modsVar = "";
@@ -172,7 +172,7 @@ namespace TaskLayer
                  raw_data = ".mzML",
                  sample_enzyme = new pepXML.Generated.msms_pipeline_analysisMsms_run_summarySample_enzyme()
                  {
-                     name = CommonParameters.Protease.Name,
+                     name = CommonParameters.DigestionParams.Protease.Name,
                      specificity = new pepXML.Generated.msms_pipeline_analysisMsms_run_summarySample_enzymeSpecificity[1]
                      {
                          new pepXML.Generated.msms_pipeline_analysisMsms_run_summarySample_enzymeSpecificity
@@ -200,8 +200,8 @@ namespace TaskLayer
                          },
                          enzymatic_search_constraint = new pepXML.Generated.msms_pipeline_analysisMsms_run_summarySearch_summaryEnzymatic_search_constraint
                          {
-                             enzyme = CommonParameters.Protease.Name,
-                             max_num_internal_cleavages = CommonParameters.MaxMissedCleavages.ToString(),
+                             enzyme = CommonParameters.DigestionParams.Protease.Name,
+                             max_num_internal_cleavages = CommonParameters.DigestionParams.MaxMissedCleavages.ToString(),
                              //min_number_termini = "2"
                          },
                          parameter = new pepXML.Generated.nameValueType[]
@@ -220,12 +220,12 @@ namespace TaskLayer
                              new pepXML.Generated.nameValueType{ name = "Cross-linker xl site", value = crosslinker.CrosslinkerModSite.ToString() },
 
                              new pepXML.Generated.nameValueType{ name = "Generate decoy proteins", value = XlSearchParameters.SearchDecoy.ToString() },
-                             new pepXML.Generated.nameValueType{ name = "MaxMissed Cleavages", value = CommonParameters.MaxMissedCleavages.ToString() },
-                             new pepXML.Generated.nameValueType{ name = "Protease", value = CommonParameters.Protease.Name },
-                             new pepXML.Generated.nameValueType{ name = "Initiator Methionine", value = CommonParameters.InitiatorMethionineBehavior.ToString() },
-                             new pepXML.Generated.nameValueType{ name = "Max Modification Isoforms", value = CommonParameters.MaxModificationIsoforms.ToString() },
-                             new pepXML.Generated.nameValueType{ name = "Min Peptide Len", value = CommonParameters.MinPeptideLength.ToString() },
-                             new pepXML.Generated.nameValueType{ name = "Max Peptide Len", value = CommonParameters.MaxPeptideLength.ToString() },
+                             new pepXML.Generated.nameValueType{ name = "MaxMissed Cleavages", value = CommonParameters.DigestionParams.MaxMissedCleavages.ToString() },
+                             new pepXML.Generated.nameValueType{ name = "Protease", value = CommonParameters.DigestionParams.Protease.Name },
+                             new pepXML.Generated.nameValueType{ name = "Initiator Methionine", value = CommonParameters.DigestionParams.InitiatorMethionineBehavior.ToString() },
+                             new pepXML.Generated.nameValueType{ name = "Max Modification Isoforms", value = CommonParameters.DigestionParams.MaxModificationIsoforms.ToString() },
+                             new pepXML.Generated.nameValueType{ name = "Min Peptide Len", value = CommonParameters.DigestionParams.MinPeptideLength.ToString() },
+                             new pepXML.Generated.nameValueType{ name = "Max Peptide Len", value = CommonParameters.DigestionParams.MaxPeptideLength.ToString() },
                              new pepXML.Generated.nameValueType{ name = "Product Mass Tolerance", value = CommonParameters.ProductMassTolerance.ToString() },
                              new pepXML.Generated.nameValueType{ name = "Ions to search", value = "B "+ CommonParameters.BIons.ToString() + " Y " + CommonParameters.YIons.ToString() + " C " + CommonParameters.CIons.ToString() + " Z " + CommonParameters.ZdotIons.ToString() },
                              new pepXML.Generated.nameValueType{ name = "Allowed Beta Precusor Mass Difference", value = XlSearchParameters.XlBetaPrecusorMsTl.ToString()},
